(#265) - return immediately when using opts.continuous
This is mostly to make the Fauxton UI nicer to work with.
As-is, it's not guaranteed to immediately show the
"replication started" message if you use continuous replication.
This fixes that.

A big TODO is to actually use the `_replicator` database,
so that you can see live replication tasks that are in-progress.